Don't perform mkfs on an argument that starts with a dash
This prevents errors when users try to call the script with arguments
like -h or --help and will instead print the usage. I think it is
reasonable to assume that a user will not want to call mkfs on a file,
or even have a file, whose name starts with a dash.