Fix calculation of filecount in statfs
The current implementation wrongly sets the f_files member to the number
of used inodes. However, f_files is supposed to be set to the maximum
number of files the filesystem supports.[1]

This leads to some bizarre output in programs that utilize statfs() like
df, which prints negative values of "inodes used".