Change root inode number from 0 to 1
This is required to stay compatible with userspace programs that avoid 0
as an inode number and fixes the odd bug that the . and .. links are not
properly displayed for the root directory even though they are emitted
correctly. This fix "wastes" one extra inode for the sake of
compatibility and thus ensures that the existing boundary checks for the
maximum inode number stay correct.