properly support truncation
Support the truncation of files to arbitrary sizes (up to the
maximum file size) by implementing ouichefs_setattr().

When the file size decreases, excess blocks are freed. When the file size
increases, the last bytes block in which the old end of file lied are zeroed,
but no new blocks are allocated, which means that the index file may contain
gaps/holes. Only when a real write to one of the new zeroed pages
is performed, new blocks will be allocated for it, which matches
the truncation implementation in most other file systems.